src/ML_app/requirements.txt
bson
paho-mqtt
azure-iot-device